Move gen_unite_graph to bm_join_plot and make it use bm-visualize
Right now, gen_unite_graph is limited to just one type of graph.

Extend it by re-using the functions inside bm-visualize. To do
that, move it to bm-runner directory.
